Style the display component
