+++ src/components/display/Display.module.css
@@ -1,0 +1,14 @@
+.display {
+  background-color: #858193;
+  color: #fff;
+  font-weight: bold;
+  padding: 0 31px;
+  height: 100px;
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+}
+
+.result {
+  font-size: 48px;
+}
